+++ ansible_mitogen/strategy.py
@@ -107,6 +107,7 @@ REDIRECTED_CONNECTION_PLUGINS = (
     'lxc',
     'lxd',
     'machinectl',
+    'podman',
     'setns',
     'ssh',
 )